+++ Arcadia/Views/Settings/CustomizeUIView.swift
@@ -1,6 +1,6 @@
 //
 //  CustomizeUIView.swift
-//  Altea
+//  Arcadia
 //
 //  Created by Davide Andreoli on 06/10/24.
 //